Add API folder and move async thunks to it
